Make PersistentLoginService interact directly with repositories
There's a problem of an unresolvable circular reference if PersistentLoginService (in general, any implementation of LoginService) tries to invoke a method that has been annotated with any authorization expression. That's because for that 'authorizing process' to succeed, a methodSecurityInterceptor has to be created first, which, in turn, indirectly depends on LoginService.